kvstore: Move to a separate package
